# openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Identity Manager GE
  description: >-
    Identity Manager (IdM) GE API specifications comply with existing standards
    for authentication and user and provide access information.
  version: '1.0'
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: '/api/v1'
paths:
...
  /users:
    get:
      tags:
        - user
      summary: List Users
      operationId: ListUsers
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
      security:
        - auth: [ ]
    post:
      tags:
        - user
      summary: Create a User
      operationId: CreateUser
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAUserRequest'
        required: true
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAUserResponse'
      security:
        - auth: [ ]
  '/users/:user_id':
    get:
      tags:
        - user
      summary: Read info about a User
      operationId: ReadUserInfo
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadInfoAboutAUserResponse'
      security:
        - auth: [ ]
    delete:
      tags:
        - user
      summary: Delete a User
      operationId: DeleteUser
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '204':
          description: ''
          content: { }
      security:
        - auth: [ ]
    patch:
      tags:
        - user
      summary: Update a User
      operationId: UpdateUser
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAUserRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAUserResponse'
      security:
        - auth: [ ]
...
  '/applications/:application_id/roles':
    get:
      tags:
        - application
      summary: List Roles
      operationId: ListRoles
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRolesResponse'
      security:
        - auth: [ ]
    post:
      tags:
        - application
      summary: Create a Role
      operationId: CreateRole
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateARoleRequest'
        required: true
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateARoleResponse'
      security:
        - auth: [ ]
  '/applications/:application_id/roles/:role_id':
    get:
      tags:
        - application
      summary: Read Role details
      operationId: ReadRoleDetails
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadRoleDetailsResponse'
      security:
        - auth: [ ]
    delete:
      tags:
        - application
      summary: Delete a Role
      operationId: DeleteRole
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '204':
          description: ''
          content: { }
      security:
        - auth: [ ]
    patch:
      tags:
        - application
      summary: Update a Role
      operationId: UpdateRole
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateARoleRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateARoleResponse'
      security:
        - auth: [ ]
...
  '/applications/:application_id/users':
    get:
      tags:
        - application
      summary: List users in an application
      operationId: ListUsersInApplication
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersInAnApplicationResponse'
      security:
        - auth: [ ]
  '/applications/:application_id/users/:user_id/roles':
    get:
      tags:
        - application
      summary: List users role assignments
      operationId: ListUsersRoleAssignments
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersRoleAssignmentsResponse'
      security:
        - auth: [ ]
  '/applications/:application_id/users/:user_id/roles/:role_id':
    put:
      tags:
        - application
      summary: Assign a role to a user
      operationId: AssignRole
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignARoleToAUserResponse'
      security:
        - auth: [ ]
    delete:
      tags:
        - application
      summary: Remove a role assignment from a user
      operationId: RemoveRoleAssignment
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '204':
          description: ''
          content: { }
      security:
        - auth: [ ]
...
  '/applications/:application_id/organizations/:organization_id/roles':
    get:
      tags:
        - application
      summary: List organizations role assignments
      operationId: ListRoleAssignments
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrganizationsRoleAssignmentsResponse'
      security:
        - auth: [ ]
...
  '/organizations/:organization_id/users':
    get:
      tags:
        - organization
      summary: List members
      operationId: ListMembers
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMembersResponse'
      security:
        - auth: [ ]
  '/organizations/:organization_id/users/:user_id/organization_roles':
    get:
      tags:
        - organization
      summary: Info of user organization relationship
      operationId: InfoOfUserOrganizationRelationship
      parameters:
        - name: X-Auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: >-
                  #/components/schemas/InfoOfUserOrganizationRelationshipResponse
      security:
        - auth: [ ]
...
components:
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        oauth_client_id:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        date_password:
          type: string
        enabled:
          type: boolean
        admin:
          type: boolean
...
    ListUsersResponse:
      title: List Users response
      required:
        - users
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
    CreateAUserRequest:
      title: Create a User request
      required:
        - user
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    CreateAUserResponse:
      title: Create a User response
      required:
        - user
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    ReadInfoAboutAUserResponse:
      title: Read info about a User response
      required:
        - user
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    UpdateAUserRequest:
      title: Update a User request
      required:
        - user
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    UpdateAUserResponse:
      title: Update a User response
      required:
        - values_updated
      type: object
      properties:
        values_updated:
          $ref: '#/components/schemas/User'
...
    ListRolesResponse:
      title: List Roles response
      required:
        - roles
      type: object
      properties:
        roles:
          type: array
          description: ''
          items:
            $ref: '#/components/schemas/Role'
    CreateARoleRequest:
      title: Create a Role request
      required:
        - role
      type: object
      properties:
        role:
          $ref: '#/components/schemas/Role'
    CreateARoleResponse:
      title: Create a Role response
      required:
        - role
      type: object
      properties:
        role:
          $ref: '#/components/schemas/Role'
    ReadRoleDetailsResponse:
      title: Read Role details response
      required:
        - role
      type: object
      properties:
        role:
          $ref: '#/components/schemas/Role'
    UpdateARoleRequest:
      title: Update a Role request
      required:
        - role
      type: object
      properties:
        role:
          $ref: '#/components/schemas/Role'
    UpdateARoleResponse:
      title: Update a Role response
      required:
        - values_updated
      type: object
      properties:
        values_updated:
          $ref: '#/components/schemas/Role'
...
    ListUsersInAnApplicationResponse:
      title: List users in an application response
      required:
        - role_user_assignments
      type: object
      properties:
        role_user_assignments:
          type: array
          description: ''
          items:
            type: object
            properties:
              user_id:
                type: string
              role_id:
                type: string
    ListUsersRoleAssignmentsResponse:
      title: List users role assignments response
      required:
        - role_user_assignments
      type: object
      properties:
        role_user_assignments:
          type: array
          description: ''
          items:
            type: object
            properties:
              user_id:
                type: string
              role_id:
                type: string
    AssignARoleToAUserResponse:
      title: Assign a role to a user response
      required:
        - role_user_assignments
      type: object
      properties:
        role_user_assignments:
          type: object
          properties:
            role_id:
              type: string
            user_id:
              type: string
            oauth_client_id:
              type: string
          description: ''
...
    ListOrganizationsRoleAssignmentsResponse:
      title: List organizations role assignments response
      required:
        - role_organization_assignments
      type: object
      properties:
        role_organization_assignments:
          type: array
          description: ''
          items:
            type: object
            properties:
              organization_id:
                type: string
              role_id:
                type: string
...
    ListMembersResponse:
      title: List members response
      required:
        - organization_users
      type: object
      properties:
        organization_users:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationUser'
    OrganizationUser:
      type: object
      properties:
        user_id:
          type: string
        organization_id:
          type: string
        role:
          type: string
    InfoOfUserOrganizationRelationshipResponse:
      title: Info of user organization relationship response
      required:
        - organization_user
      type: object
      properties:
        organization_user:
          type: object
          description: ''
          properties:
            user_id:
              type: string
            organization_id:
              type: string
            role:
              type: string
...
  securitySchemes:
    auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'https://api.vixal.xyz/api'
          scopes: { }